Started work on BVH.

+++ src/util.h
@@ -5,7 +5,7 @@
 
 namespace mcp
 {
-    static const float kEpsilon = 1e-6;
+    static const double kEpsilon = 1e-6;
 
     template <typename T>
     inline T clamp(T value, T min, T max) {
